requirement.txt
streamlit
pillow
supabase
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas